refactor(data/remotesociallogin): move files to your folder
