+++ src/data/errors/index.ts
@@ -3,3 +3,4 @@ export { default as UnexpectedError } from './unexpectedError';
 export { default as UserNotFoundError } from './userNotFoundError';
 export { default as UserNotHaveAccessError } from './userNotHaveAccessError';
 export { default as NoAccessError } from './noAccessError';
+export { default as SocialLoginError } from './socialLoginError';
