+++ src/data/useCases/index.ts
@@ -4,3 +4,4 @@ export { default as RemoteGetUser } from './remoteGetUser';
 export { default as RemoteListFavoritePlaces } from './remoteListFavoritePlaces';
 export { default as RemoteListRoadmap } from './remoteListRoadmap';
 export { default as RemoteSearchPlaces } from './remoteSearchPlaces';
+export { default as RemoteSocialLogin } from './remoteSocialLogin';
