+++ src/domain/useCases/index.ts
@@ -4,3 +4,4 @@ export { default as ListFavoritePlaces } from './listFavoritePlaces';
 export { default as ListRoadmap } from './listRoadmap';
 export { default as SearchPlaces } from './searchPlaces';
 export { default as ListPlaces } from './listPlaces';
+export { default as SocialLogin } from './socialLogin';
